grpc_backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
exceptiongroup==1.3.1
fastapi==0.123.5
grpcio==1.76.0
grpcio-tools==1.76.0
idna==3.11
protobuf==6.33.1
pydantic==2.12.5
pydantic_core==2.41.5
SQLAlchemy==2.0.44
sqlmodel==0.0.27
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
